+++ lib/formex/utils.ex
@@ -23,27 +23,3 @@ defmodule Formex.Utils do
   end
 
 end
-
-defmodule Formex.Utils.Counter do
-  use GenServer
-
-  def start_link do
-    GenServer.start_link(__MODULE__, 0)
-  end
-
-  def increment(pid) do
-    GenServer.call(pid, :increment)
-  end
-
-  def reset(pid) do
-    GenServer.cast(pid, :reset)
-  end
-
-  def handle_call(:increment, _from, state) do
-    {:reply, state, state + 1}
-  end
-
-  def handle_cast(:reset, _state) do
-    {:noreply, 0}
-  end
-end